requirements.txt
pysam==0.11
synbiochem-py